Move router to main source directory A seperate folder is unnecessary for just a single file
