+++ src/main.ts
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import App from "./App.vue";
-import router from "./router/router";
+import router from "./router";
 import store from "./store/store";
 import vuetify from "./plugins/vuetify";
